add new main function
this function carry out the statistic result of the trade volume during
the last few mins every trade day.
The generated figures display the probability distribution of
trade_volume_last_few_mins/trade_volume_whole_day, which indicates the
portion of the trade volume during the last few mins to the volume of
the whole day.